图像分类-单机/requirements.txt
jieba
h5py
keras
traitlets
ipython
pytest
dask
psutil
ipywidgets
paramiko
seaborn
ray
matplotlib
tqdm
numpy
pandas
scikit-learn
pillow
opencv-python
shap
scipy
fsspec
joblib
setuptools